@@ -13,26 +13,24 @@
         {"name": "Johannes Schmitt", "email": "schmittjoh@gmail.com"}
     ],
     "require": {
-        "php": ">=5.3.2",
+        "php": "~5.5|~7.0",
         "doctrine/inflector": "1.*",
         "doctrine/cache": "1.*",
         "doctrine/collections": "1.*",
         "doctrine/lexer": "1.*",
         "doctrine/annotations": "1.*"
     },
-    "minimum-stability": "dev",
     "require-dev": {
-        "phpunit/phpunit": "~3.7"
+        "phpunit/phpunit": "~4.8|~5.0"
     },
     "autoload": {
-        "psr-0": { "Doctrine\\Common\\": "lib/" }
+        "psr-4": {
+            "Doctrine\\Common\\": "lib/Doctrine/Common"
+        }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.5.x-dev"
+            "dev-master": "2.7.x-dev"
         }
-    },
-    "archive": {
-        "exclude": ["!vendor", "tests", "*phpunit.xml", ".travis.yml", "build.xml", "build.properties", "composer.phar"]
     }
 }
