@@ -11,16 +11,20 @@
         {"name": "Benjamin Eberlei", "email": "kontakt@beberlei.de"},
         {"name": "Jonathan Wage", "email": "jonwage@gmail.com"}
     ],
+    "minimum-stability": "dev",
     "require": {
-        "php": ">=5.3.2",
+        "php": ">=5.4",
         "ext-pdo": "*",
-        "doctrine/collections": "~1.1",
-        "doctrine/dbal": "~2.4",
-        "symfony/console": "~2.0"
+        "doctrine/collections": "~1.2",
+        "doctrine/dbal": ">=2.5-dev,<2.6-dev",
+        "doctrine/instantiator": "~1.0.1",
+        "doctrine/common": ">=2.5-dev,<2.7-dev",
+        "doctrine/cache": "~1.4",
+        "symfony/console": "~2.5|~3.0"
     },
     "require-dev": {
-        "symfony/yaml": "~2.1",
-        "satooshi/php-coveralls": "dev-master"
+        "symfony/yaml": "~2.3|~3.0",
+        "phpunit/phpunit": "~4.0"
     },
     "suggest": {
         "symfony/yaml": "If you want to use YAML Metadata Mapping Driver"
@@ -28,10 +32,13 @@
     "autoload": {
         "psr-0": { "Doctrine\\ORM\\": "lib/" }
     },
+    "autoload-dev": {
+        "psr-0": { "Doctrine\\Tests\\": "tests/" }
+    },
     "bin": ["bin/doctrine", "bin/doctrine.php"],
     "extra": {
         "branch-alias": {
-            "dev-master": "2.4.x-dev"
+            "dev-master": "2.6.x-dev"
         }
     },
     "archive": {
